rewrote the affinity propagation sequence to a) be correct, b) stay mostly on the gpu with fewer downloads. still quite slow, I think having such a long sequence of iteration within the kernel is bad. I have an idea to make the whole process much more efficient.

+++ Plumbus/Cuda_functions/headers/form_similarity_matrix_color.cuh
@@ -5,4 +5,4 @@
 
 
 
-void form_similarity_matrix_color_launch(thrust::device_vector<int> &source, cv::cuda::GpuMat &output, int N);
+void form_similarity_matrix_color_launch(thrust::device_vector<float> &source, cv::cuda::GpuMat &output, int N);
